pyproject.toml
[tool.poetry]
name = "lark_shell"
version = "0.1.0"
description = "A terminal Lark IDE"
authors = ["Bryan Hu <bryan.hu.2020@gmail.com>"]
license = "GNU GPL v3+"

[tool.poetry.dependencies]
python = "^3.6"
urwid = "^2.1.2"
lark-parser = "^0.10.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
pylint = "^2.6.0"
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
